requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyodbc==5.1.0
httpx==0.27.2
python-dotenv==1.0.1
email-validator==2.2.0
